Replacing UIKIt app lifecycle with SwiftUI approach.  Also put Core Data stack in its own struct.
