+++ DhammaTalks/CoreDataStack.swift
@@ -1,0 +1,40 @@
+//
+//  CoreDataStack.swift
+//  DhammaTalks
+//
+//  Created by Bill Parousis on 2021-12-18.
+//  Copyright © 2021 Bill Parousis. All rights reserved.
+//
+
+import CoreData
+
+struct CoreDataStack {
+    static var viewContext: NSManagedObjectContext {
+        return persistentContainer.viewContext
+    }
+    
+    static var persistentContainer: NSPersistentContainer = {
+        let container = NSPersistentContainer(name: "DhammaTalks")
+        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+            container.viewContext.mergePolicy = NSMergeByPropertyObjectTrumpMergePolicy
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        return container
+    }()
+
+    static func saveContext() {
+        let context = persistentContainer.viewContext
+        if context.hasChanges {
+            do {
+                try context.save()
+            } catch {
+                // Replace this implementation with code to handle the error appropriately.
+                // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
+                let nserror = error as NSError
+                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+            }
+        }
+    }
+}
