+++ DhammaTalks/DhammaTalksApp.swift
@@ -1,0 +1,32 @@
+//
+//  DhammaTalksApp.swift
+//  DhammaTalks
+//
+//  Created by Bill Parousis on 2021-12-18.
+//  Copyright © 2021 Bill Parousis. All rights reserved.
+//
+
+import SwiftUI
+import UIKit
+import CoreData
+
+@main
+struct DhammaTalksApp: App {
+    
+    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+    private let talkUserInfoService: TalkUserInfoService
+    private let talkDataService: TalkDataService
+    
+    init() {
+        self.talkUserInfoService = TalkUserInfoService(managedObjectContext: CoreDataStack.viewContext)
+        self.talkDataService = TalkDataService()
+    }
+
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+                .environmentObject(talkUserInfoService)
+                .environmentObject(talkDataService)
+        }
+    }
+}
